Feature:  Support multiple versions of ClientTools
In conjunction with "multiple target platform" support, the user may
also have several versions of the Client Tools packages.

The goal is to allow the user to arbitrarily specify targets and client
tools, but continue to work with a single ECL code base.

Note:  Should the versions diverge to such an extent that the ECL could
never be compatible between them, it will be recommended to use the
Eclipse "Workspace" feature to separate them out (IOW the user has the
choice on how to configure their environment appropriately).

Signed-off-by: Gordon Smith <gordon.smith@lexisnexis.com>